requirements.txt
django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.2
python-dotenv>=1.0
dj-rest-auth>=3.2
djoser>=2.1
gunicorn
django-cors-headers>=3.15
cloudinary>=1.30
django-cloudinary-storage>=0.3.0
django-filter>=23.0
